IntelligentSearchAgent/requirements.txt
google-adk
requests
beautifulsoup4
tavily-python
python-dotenv